@@ -20,7 +20,7 @@
       "exclude": ["/Tests"]
     },
     "require": {
-        "php":                                 ">=5.4",
+        "php":                                 "^5.6|^7.0",
         "symfony/framework-bundle":            "~2.4|~3.0",
         "symfony/twig-bundle":                 "~2.4|~3.0",
         "symfony/form":                        "~2.4|~3.0",
@@ -33,7 +33,8 @@
         "matthiasnoback/symfony-config-test":  "~1",
         "matthiasnoback/symfony-dependency-injection-test": "~0.7",
         "symfony/finder": "~2.4|3.0",
-        "satooshi/php-coveralls": "~1"
+        "satooshi/php-coveralls": "~1",
+        "phpunit/phpunit": "~5"
     },
     "suggest": {
         "helios-ag/fm-tinymce-bundle": "TinyMCE Bundle by me",
