@@ -12,7 +12,7 @@ use Symfony\Component\Form\FormView;
  *
  * @package FM\ElfinderBundle\Tests
  */
-class ElFinderTypeTest extends \PHPUnit_Framework_TestCase
+class ElFinderTypeTest extends \PHPUnit\Framework\TestCase
 {
     public function testGetName()
     {
@@ -64,7 +64,7 @@ class ElFinderTypeTest extends \PHPUnit_Framework_TestCase
         );
         $view = new FormView();
         $type = new ElFinderType();
-        $form = $this->getMock('Symfony\Component\Form\Test\FormInterface');
+        $form = $this->createMock('Symfony\Component\Form\Test\FormInterface');
         $type->buildView($view, $form, $options);
         foreach ($options as $name => $value) {
             $this->assertArrayHasKey($name, $view->vars);
